backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
openai==0.27.6
langchain==0.0.150
python-dotenv==1.0.0
scikit-learn==1.2.2
PyPDF2==3.0.1
opencv-python-headless==4.7.0.72
pytesseract==0.3.9
Pillow==9.5.0
faiss-cpu==1.7.4
python-multipart==0.0.6



# fastapi==0.109.0
# uvicorn==0.27.1
# langchain-community==0.0.29
# faiss-cpu==1.7.4
# langchain-openai>=0.1.0
# openai>=1.12.0
# python-multipart==0.0.9
# pydantic==2.6.1
# python-dotenv==1.0.0
# requests==2.31.0
# tiktoken==0.5.2
# unstructured==0.12.2
